Fixed a crash when casting GN_SPORE_EXPLOSION on certain systems.
- Follow-up to 44c33fda3614d588e6bf6cee1cf884e98f1531f0.
- Also fixed some incorrect va_end calls that came up while debugging.
- Special thanks to Ind for his help debugging the problem.

+++ src/map/skill.c
@@ -3198,7 +3198,7 @@ int skill_timerskill(int tid, int64 tick, int id, intptr_t data) {
 					break;
 				case GN_SPORE_EXPLOSION:
 					map->foreachinrange(skill->area_sub, target, skill->get_splash(skl->skill_id, skl->skill_lv), BL_CHAR,
-					                    src, skl->skill_id, skl->skill_lv, 0, skl->flag|1|BCT_ENEMY, skill->castend_damage_id);
+					                    src, skl->skill_id, skl->skill_lv, (int64)0, skl->flag|1|BCT_ENEMY, skill->castend_damage_id);
 					break;
 				case SR_FLASHCOMBO_ATK_STEP1:
 				case SR_FLASHCOMBO_ATK_STEP2:
